Gen4: moved gen4 select plan to selectGen4 and renamed all code from v4 to Gen4
