+++ go/vt/vtgate/planbuilder/pullout_subquery.go
@@ -88,11 +88,11 @@ func (ps *pulloutSubquery) Wireup(plan logicalPlan, jt *jointab) error {
 }
 
 // Wireup2 implements the logicalPlan interface
-func (ps *pulloutSubquery) WireupV4(semTable *semantics.SemTable) error {
-	if err := ps.underlying.WireupV4(semTable); err != nil {
+func (ps *pulloutSubquery) WireupGen4(semTable *semantics.SemTable) error {
+	if err := ps.underlying.WireupGen4(semTable); err != nil {
 		return err
 	}
-	return ps.subquery.WireupV4(semTable)
+	return ps.subquery.WireupGen4(semTable)
 }
 
 // SupplyVar implements the logicalPlan interface
