+++ go/vt/vtgate/planbuilder/sql_calc_found_rows.go
@@ -41,12 +41,12 @@ func (s *sqlCalcFoundRows) Wireup(logicalPlan, *jointab) error {
 }
 
 //Wireup2 implements the logicalPlan interface
-func (s *sqlCalcFoundRows) WireupV4(semTable *semantics.SemTable) error {
-	err := s.LimitQuery.WireupV4(semTable)
+func (s *sqlCalcFoundRows) WireupGen4(semTable *semantics.SemTable) error {
+	err := s.LimitQuery.WireupGen4(semTable)
 	if err != nil {
 		return err
 	}
-	return s.CountQuery.WireupV4(semTable)
+	return s.CountQuery.WireupGen4(semTable)
 }
 
 // Solves implements the logicalPlan interface
